adjusted sqlite behavior if compressed file already detected

+++ Compressor/compressor.fs
@@ -3,6 +3,7 @@
 open System
 open System.IO
 open System.IO.Compression
+open Compressor.log
 open Compressor.compress
 open Compressor.types
 open Microsoft.FSharp.Collections
@@ -39,33 +40,34 @@ let private collect_data_files path extension cutoff =
 // crude
 let private compress_file file remove_after =
         let compressed_path = $"{file}.gz"
-        use input_file_stream = File.OpenRead(file)
-        use output_file_stream = File.Create(compressed_path)
-        use gz_stream = new GZipStream(output_file_stream, CompressionLevel.Optimal)
-
-        try
-            input_file_stream.CopyTo(gz_stream)
-            input_file_stream.Flush()
-            input_file_stream.Close()
-            gz_stream.Flush() // need to flush to get the size
-            gz_stream.Close()
-            match remove_after with
-            | true -> File.Delete(file)
-            | false -> ()
-
-            let bytes_saved = FileInfo(compressed_path).Length
-            printfn $"bytes_saved : {bytes_saved}"
-            Success {
-                Date = DateTime.Now
-                CompressedFilename = compressed_path
-                Removed = remove_after
-                BytesSaved = bytes_saved
-            }
-
-        with
-            | _ as ex -> 
-                File.Delete(compressed_path)
-                Failure { Date = DateTime.Now; Filename = file; Error = $"Failed to compress {file} : {ex.Message}" }
+        if File.Exists(compressed_path) then
+            let name = FileInfo(file).Name
+            Failure { Date = DateTime.Now; Filename = file; Error = $"{name} already compressed!" }
+        else
+            use input_file_stream = File.OpenRead(file)
+            use output_file_stream = File.Create(compressed_path)
+            use gz_stream = new GZipStream(output_file_stream, CompressionLevel.Optimal)
+            try
+                input_file_stream.CopyTo(gz_stream)
+                input_file_stream.Flush()
+                input_file_stream.Close()
+                gz_stream.Flush() // need to flush to get the size
+                gz_stream.Close()
+                match remove_after with
+                | true -> File.Delete(file)
+                | false -> ()
+                let bytes_saved = FileInfo(compressed_path).Length
+                printfn $"bytes_saved : {bytes_saved}"
+                Success {
+                    Date = DateTime.Now
+                    CompressedFilename = compressed_path
+                    Removed = remove_after
+                    BytesSaved = bytes_saved
+                }
+            with
+                | _ as ex -> 
+                    File.Delete(compressed_path)
+                    Failure { Date = DateTime.Now; Filename = file; Error = $"Failed to compress {file} : {ex.Message}" }
 
 
 let compress_dir config =
@@ -77,7 +79,6 @@ let compress_dir config =
                 let! result = compress_file file config.RemoveAfter
                 return result
             })
-
     let mb_saved = 
         results 
         |> Seq.choose (fun comp ->
@@ -86,5 +87,4 @@ let compress_dir config =
             | Failure _ -> None
         )
         |> Seq.fold (fun acc compressed -> (float compressed.BytesSaved / 1_000_000.) + acc) 0.
-
     mb_saved
